Update Tile Movement Game.py
I got rid of the attempt at getting the sword to spawn on the side of the player relative to the direction that the player is moving. The sword now spawn well on the right side of the player. I have implemented a 'flag' system with "swordavaliable" to ensure that the player does not spawn infinite sword for the duration of the SPACE key being held down - this all works well